+++ src/HighScore/HighScore.h
@@ -7,14 +7,14 @@
 class HighScore {
  public:
     static bool latestScoreIsOnLeaderBoard;
-    static int latestPlaceOnLeaderBoard;
+    static uint8_t latestPlaceOnLeaderBoard;
     static uint16_t highScores[5];
 
     static void init();
     static void newScore(uint16_t score);
     static uint16_t* getLeaderBoard();
     static bool isLatestScoreOnLeaderBoard();
-    static int getPlaceOnLeaderBoard();
+    static uint8_t getPlaceOnLeaderBoard();
     static void resetHighScores();
 
  private:
